Update Dusk2Dawn.h
Actually, it is not possible to calculate dusk and dawn for fractional timezones.
This because the main constructor (Dusk2Dawn) is correctly accepting fractional timezones, but the variable that keeps the timezone (_timezone) is an integer.
To fix this, it is enough to change the type of the variable from integer to float.

+++ Dusk2Dawn.h
@@ -18,7 +18,7 @@
       static bool min2str(char*, int);
     private:
       float _latitude, _longitude;
-      int   _timezone;
+      float    _timezone;
       int   sunriseSet(bool, int, int, int, bool);
       float sunriseSetUTC(bool, float, float, float);
       float equationOfTime(float);
